Organize centralreport class.
