+++ centralreport/centralreport.py
@@ -3,15 +3,22 @@
 # CentralReport - Indev version
 # Project by Charles-Emmanuel CAMUS - Avril 2012
 
-import utils.CRLog, utils.CRConfig
+import sys,time,utils.CRLog, utils.CRConfig
+from daemon import Daemon
 from utils.CRConfig import CRConfig
 from threads.ThreadChecks import ThreadChecks
 from web.webserver import WebServer
 
-class CentralReport():
+class CentralReport(Daemon):
 
-    def __init__(self):
+    # Bool : True = daemon is running.
+    isRunning = True
+
+    def run(self):
         # Constructeur
+
+        # False = no error
+        isError = False
 
         # On prepare les logs
         utils.CRLog.CRLog.configLog()
@@ -33,11 +40,12 @@ class CentralReport():
             ThreadChecks()
 
         else:
+            isError = True
             print("Sorry, but your OS is not supported yet...")
 
 
         # Enable webserver ?
-        if CRConfig.config_webserver_enable == True:
+        if (CRConfig.config_webserver_enable == True) & (isError == False):
             # Yeah !
             print("Enabling the webserver")
             #WebServer().start()
@@ -47,11 +55,15 @@ class CentralReport():
             print("Webserver is disabled by configuration file")
             utils.CRLog.CRLog.writeInfo("Webserver is disabled by configuration file")
 
-
-
-
         # End of file
-        utils.CRLog.CRLog.writeInfo("CentralReport started!")
+        if(isError == False):
+            utils.CRLog.CRLog.writeInfo("CentralReport started!")
+
+            while self.isRunning:
+                time.sleep(1)
+
+        else:
+            utils.CRLog.CRLog.writeError("Error launching CentralReport!")
 
 
     def stop(self):
@@ -59,4 +71,36 @@ class CentralReport():
         Called when the scripts will be stopped
         """
 
-        utils.CRLog.CRLog.writeInfo("Stopping CentralReport")
+        utils.CRLog.CRLog.writeInfo("Stopping CentralReport...")
+
+        self.isRunning = False
+        Daemon.stop(self)
+
+
+#
+# Main script
+#
+
+if __name__ == "__main__":
+
+    # Launching the daemon...
+    daemon = CentralReport(utils.CRConfig.CRConfig.pid_file)
+
+    if len(sys.argv) == 2:
+        if 'start' == sys.argv[1]:
+            print("CentralReport -- Start")
+            daemon.start()
+        elif 'stop' == sys.argv[1]:
+            daemon.stop()
+            print("CentralReport -- Stopped")
+        elif 'restart' == sys.argv[1]:
+            print ("CentralReport -- Restarting...")
+            daemon.restart()
+            print ("CentralReport -- Started")
+        else:
+            print "Unknown command"
+            sys.exit(2)
+        sys.exit(0)
+    else:
+        print "usage: %s start|stop|restart" % sys.argv[0]
+        sys.exit(2)
