+++ centralreport/config.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+
+#
+# CentralReport - Indev version
+#
+
+
+print '\n--------------------------------------------------'
+print '         CentralReport config editor'
+print '--------------------------------------------------'
+print '\nYou can also edit manually the config file, located at /etc/cr/centralreport.cfg'
+
+print '\nChecking if CentralReport is started...'
+
+
