+++ launchers/com.centralreport.plist
@@ -16,7 +16,7 @@
         <key>ProgramArguments</key>
         <array>
                 <string>python</string>
-                <string>/usr/local/bin/centralreport/run.py</string>
+                <string>/usr/local/bin/centralreport/centralreport.py</string>
                 <string>start</string>
         </array>
 </dict>
